Avoid call to strlen in fast_c_string_match_ignore_case. * search.c (fast_c_string_match_ignore_case): Change to use length argument.  Adjust users accordingly. * lisp.h (fast_c_string_match_ignore_case): Adjust prototype.

+++ src/font.h
@@ -782,8 +782,8 @@ extern void font_parse_family_registry (Lisp_Object family,
                                         Lisp_Object spec);
 
 extern int font_parse_xlfd (char *name, ptrdiff_t len, Lisp_Object font);
-extern int font_unparse_xlfd (Lisp_Object font, int pixel_size,
-                              char *name, int bytes);
+extern ptrdiff_t font_unparse_xlfd (Lisp_Object font, int pixel_size,
+				    char *name, int bytes);
 extern int font_unparse_fcname (Lisp_Object font, int pixel_size,
                                 char *name, int bytes);
 extern void register_font_driver (struct font_driver *driver, FRAME_PTR f);
